CONN-10463 Move MetadataRecord back to tests
